Add files via upload
In this project, I've leveraged Python's Pillow library to open an image file and convert it into a numpy array. By examining the properties and structure of this array with numpy methods, the project reveals a fundamental truth: every image is essentially a 3-dimensional array composed of RGB values. Through the manipulation of these R, G, and B components, we can create striking visual transformations, demonstrating how subtle changes to these values can dramatically alter the appearance of an image.